- 格式化doc - 增加断点续传功能（未测试

+++ src/utils/downloader/helpers.rs
@@ -4,13 +4,13 @@ use anyhow::{Context, Ok, Result};
 use reqwest::Response;
 
 /// 解析响应头中的 Content-Length 并返回 u64
-/// 
+///
 /// # Arguments
-/// 
+///
 /// * `resp` - HTTP响应的引用
-/// 
+///
 /// # Returns
-/// 
+///
 /// * `Result<u64>` - 成功时返回文件长度，失败时返回错误
 pub fn get_file_length(resp: &Response) -> Result<u64> {
     let headers = resp.headers();
@@ -48,14 +48,14 @@ pub fn get_file_length(resp: &Response) -> Result<u64> {
 }
 
 /// 同步解压GZ文件
-/// 
+///
 /// # Arguments
-/// 
+///
 /// * `gz_path` - GZ压缩文件路径的引用
 /// * `output_path` - 输出文件路径的引用
-/// 
+///
 /// # Returns
-/// 
+///
 /// * `Result<()>` - 成功时返回空值，失败时返回错误
 pub fn decompress_gz_sync(gz_path: &Path, output_path: &Path) -> Result<()> {
     use flate2::read::GzDecoder;
@@ -71,13 +71,13 @@ pub fn decompress_gz_sync(gz_path: &Path, output_path: &Path) -> Result<()> {
 }
 
 /// 异步获取URL响应
-/// 
+///
 /// # Arguments
-/// 
+///
 /// * `url` - URL字符串的引用
-/// 
+///
 /// # Returns
-/// 
+///
 /// * `Result<Response>` - 成功时返回HTTP响应，失败时返回错误
 pub async fn url_get(url: &str) -> Result<Response> {
     use super::build_request;
@@ -90,12 +90,7 @@ pub async fn url_get(url: &str) -> Result<Response> {
         .context("reqwest error in url get")
 }
 
-pub async fn url_get_range(
-    url: &str,
-    range_type: &str,
-    start: usize,
-    end: usize,
-) -> Result<Response> {
+pub async fn url_get_range(url: &str, range_type: &str, start: u64, end: u64) -> Result<Response> {
     use super::build_request;
     use super::create_http_client;
